First skeleton of Omni4Controller. Build ok,. Nothing implemented
