+++ omni4_controller/controller_plugins.xml
@@ -1,0 +1,9 @@
+<library path="lib/libmy_controller_lib">
+	<class 	name="omni4_controller/Omni4Controller"
+	     	type="omni4_controller::Omni4Controller"
+	     	base_class_type="controller_interface::ControllerBase">
+	 	<description>
+			Kinematics controller for a 4-omniwheel mobile robot
+		</description>
+ 	</class>
+</library>
