--- omni_4_controller/controller_plugins.xml
@@ -1,5 +1,0 @@
-library path="lib/libmy_controller_lib">
-  <class name="omni_4_controller/Omni4Controller"
-         type="omni_4_controller::Omni4Controller"
-         base_class_type="controller_interface::ControllerBase" />
-</library>
